Add recv_with_fds
This function is not the best that it could be.
It would be nice if it could return an arbitrary number of file descriptors,
and allow for optional file descriptors.
But it serves a purpose by making the code simpler and reducing repetition.

With this added function I think it's finally time to clean up the {send,recv}*fd[s]?
family of functions, there are just too many.

My ideal world would have:
send_with_fds (call with &mut [] if you have nothing)
send
recv_with_fds (see above)
recv

In reality recv_(with)?_fd might be kept since it is simpler than const
generic madness.