Move index.html dance from fs process to client process.
The fs process is simple and doesn't need to know about these weird ideas.

Currently is an index.html does not exist then a connection has to be setup
with the fs process twice.
To improve this case the fs process could accept multiple requests on a single
received connection.